requirements.txt
# Streamlit Cloud Optimized Dependencies
# For deployment to Streamlit Cloud - minimal but functional

# Core Web Framework
streamlit>=1.32.0
python-dotenv>=1.0.0

# Image Processing
pillow>=10.0.0

# LangChain Core (essential only)
langchain==0.1.20
langchain-core==0.1.45
langchain-openai==0.0.8
langchain-google-genai==1.0.1
langchain-community==0.0.38
langchain-chroma==0.1.1

# Vector Database
chromadb>=0.4.24

# AI Providers
openai>=1.12.0
google-generativeai>=0.4.0

# Document Processing (lightweight alternatives)
pypdf>=4.0.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Utilities
numpy>=1.24.0
requests>=2.31.0
tqdm>=4.66.0
tenacity>=8.2.0

# NLP (minimal)
nltk>=3.8.1

# Data Handling
pandas>=2.0.0

# Pydantic for data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
